Merged PR 13486: Modify syncInMemoryLocales to support upserts
In the template Catalog Service, we do not expect multiple assets to have different locale combinations for localizations. But in BeyondDutch, this is not the case.

Consider the following for example:
Movie A has localizations for `EE` and `SV`, in addition to the `default` locale.
Movie B has localizations for `EE` and `NL`, in addition to the `default` locale.

According to the template Catalog Service logic, when Movie A is published, it updates the locales table (and the inMemoryLocales array) with `EE` and `SV` locales.
Then when Movie B is published, it deletes the existing locales and adds only the locales `EE` and `NL`. This is because in template Media Service, there're validations which makes sure all published assets have the same combination of locales.

But in BeyondDutch, these locales can be different from one asset to the other. This PR modifies the in memory locales logic to support this scenario.

Related work items: #47135