Merged PR 13531: Change collation to und-x-icu in title columns
By default the Collation in all columns in all tables defaults to `en_US.UTF8`. But this collation type does not behave as expected when sorting with titles (with accent characters) and with numbers.

This PR changes the collation of title columns to `und-x-icu`, which is the recommended default by Postgres for multi-locale scenarios. (https://www.postgresql.org/docs/16/collation.html#COLLATION-MANAGING-PREDEFINED-ICU)

Related work items: #47239