Merged PR 13553: Create localization views with returning default localization if requested locale is not defined
BD does not define translations for all locales for a given asset. Because of this, when requested with a specific locale, if no translation is defined for the requested locale, it gives an empty response.

The logic for creating localization views is modified through this PR to return the localization for the `default` locale if the requested locale is not available.

Note:
Ideally, the solution should be that when publishing from the Classic CMS, if there are not localizations defined for all locales, we need to publish the default localizations for those undefined locales as well. This will simplify the view in Catalog Service, and possibly improve the performance by removing the extra joins I've introduced now.

But since we do not have a correct way of getting the list of locales systemwide in a way we can put it in the publishing code from Classic CMS, we'll go with this fix, until we find a better way forward.
FYI @<Madawa Illesinghe> @<Sachith Ranasinghe>

Related work items: #47135